Add a cache for the weather requests
Till now, every weather releated command resulted in an additional
request to the weather API.
Now there is a cache with a 15min lifespan. This means that the complete
response from the server will be saved for 15 minutes after the initial
request. In this time span every new request will get handled much
faster.

Morover, the automatic weather message is now also much cheaper. First,
because before it resulted in a new API request for every user, which a
cache would have prevented but also because now it only generates the
whole weather text only once for all users.